requirements.txt
tqdm==4.66.1
scikit-learn==1.3.2
scipy==1.14.1
matplotlib==3.8.2
pillow==10.0.1
pandas==2.1.3
seaborn==0.13.0
seaborn-image==0.7.1
numpy==1.26.2
h5py==3.10.0
openpyxl==3.1.5
jupyterlab==4.3.5
PyYAML==6.0.2
